+++ src/solvers/SolverNKA_BT.hh
@@ -51,7 +51,7 @@ class SolverNKA_BT : public Solver<Vector, VectorSpace> {
 
   // mutators
   void set_tolerance(double tol) { tol_ = tol; }
-  void set_pc_lag(double pc_lag) { pc_lag_ = pc_lag; }
+  void set_pc_lag(int pc_lag) { pc_lag_ = pc_lag; }
 
   // access
   double tolerance() { return tol_; }
